fixes #760

+++ tests/unit/relations/ManyToManyRelation.js
@@ -960,13 +960,7 @@ describe('ManyToManyRelation', () => {
         expect(executedQueries[0]).to.equal(builder.toString());
         expect(executedQueries[0]).to.equal(builder.toSql());
         expect(executedQueries[0]).to.eql(
-          [
-            'update "RelatedModel" set "a" = \'str1\'',
-            'where "RelatedModel"."rid" in',
-            '(select "JoinModel"."relatedId" from "JoinModel" where "JoinModel"."ownerId" = 666)',
-            'and "gender" = \'male\'',
-            'and "thingy" is not null'
-          ].join(' ')
+          `update "RelatedModel" set "a" = 'str1' where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "gender" = 'male' and "thingy" is not null)`
         );
       });
     });
@@ -992,14 +986,7 @@ describe('ManyToManyRelation', () => {
         expect(executedQueries[0]).to.equal(builder.toString());
         expect(executedQueries[0]).to.equal(builder.toSql());
         expect(executedQueries[0]).to.eql(
-          [
-            'update "RelatedModel" set "a" = \'str1\'',
-            'where ("RelatedModel"."cid","RelatedModel"."did") in',
-            '(select "JoinModel"."relatedCId", "JoinModel"."relatedDId" from "JoinModel"',
-            'where "JoinModel"."ownerAId" = 11 and "JoinModel"."ownerBId" = 22)',
-            'and "gender" = \'male\'',
-            'and "thingy" is not null'
-          ].join(' ')
+          `update "RelatedModel" set "a" = 'str1' where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."cid" = "JoinModel"."relatedCId" and "RelatedModel"."did" = "JoinModel"."relatedDId" where ("JoinModel"."ownerAId", "JoinModel"."ownerBId") in ((11, 22)) and "gender" = 'male' and "thingy" is not null)`
         );
       });
     });
@@ -1022,14 +1009,7 @@ describe('ManyToManyRelation', () => {
           expect(numUpdated).to.equal(42);
           expect(executedQueries).to.have.length(1);
           expect(executedQueries[0]).to.eql(
-            [
-              'update "RelatedModel" set "a" = \'str1\'',
-              'where "RelatedModel"."rid" in',
-              '(select "JoinModel"."relatedId" from "JoinModel"',
-              'where "JoinModel"."ownerId" = 666)',
-              'and "gender" = \'male\'',
-              'and "thingy" is not null'
-            ].join(' ')
+            `update "RelatedModel" set "a" = 'str1' where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "gender" = 'male' and "thingy" is not null)`
           );
         });
     });
@@ -1050,15 +1030,7 @@ describe('ManyToManyRelation', () => {
         .then(() => {
           expect(executedQueries).to.have.length(1);
           expect(executedQueries[0]).to.eql(
-            [
-              'update "RelatedModel" set "a" = \'str1\'',
-              'where "RelatedModel"."rid" in',
-              '(select "JoinModel"."relatedId" from "JoinModel"',
-              'where "JoinModel"."ownerId" = 666)',
-              'and "someColumn" = 100',
-              'and "gender" = \'male\'',
-              'and "thingy" is not null'
-            ].join(' ')
+            `update "RelatedModel" set "a" = 'str1' where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "someColumn" = 100 and "gender" = 'male' and "thingy" is not null)`
           );
         });
     });
@@ -1086,12 +1058,7 @@ describe('ManyToManyRelation', () => {
         expect(executedQueries[0]).to.equal(builder.toString());
         expect(executedQueries[0]).to.equal(builder.toSql());
         expect(executedQueries[0]).to.eql(
-          [
-            'update "RelatedModel" set "a" = \'str1\'',
-            'where "RelatedModel"."rid" in (select "JoinModel"."relatedId" from "JoinModel" where "JoinModel"."ownerId" = 666)',
-            'and "gender" = \'male\'',
-            'and "thingy" is not null'
-          ].join(' ')
+          `update "RelatedModel" set "a" = 'str1' where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "gender" = 'male' and "thingy" is not null)`
         );
       });
     });
@@ -1124,12 +1091,7 @@ describe('ManyToManyRelation', () => {
           expect(numUpdated).to.equal(42);
           expect(executedQueries).to.have.length(1);
           expect(executedQueries[0]).to.eql(
-            [
-              'update "RelatedModel" set "a" = \'str1\'',
-              'where "RelatedModel"."rid" in (select "JoinModel"."relatedId" from "JoinModel" where "JoinModel"."ownerId" = 666)',
-              'and "gender" = \'male\'',
-              'and "thingy" is not null'
-            ].join(' ')
+            `update "RelatedModel" set "a" = 'str1' where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "gender" = 'male' and "thingy" is not null)`
           );
         });
     });
@@ -1145,12 +1107,7 @@ describe('ManyToManyRelation', () => {
         .then(() => {
           expect(executedQueries).to.have.length(1);
           expect(executedQueries[0]).to.eql(
-            [
-              'update "RelatedModel" set "test" = "test" + 1',
-              'where "RelatedModel"."rid" in',
-              '(select "JoinModel"."relatedId" from "JoinModel"',
-              'where "JoinModel"."ownerId" = 666)'
-            ].join(' ')
+            `update "RelatedModel" set "test" = "test" + 1 where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666))`
           );
         });
     });
@@ -1166,12 +1123,7 @@ describe('ManyToManyRelation', () => {
         .then(() => {
           expect(executedQueries).to.have.length(1);
           expect(executedQueries[0]).to.eql(
-            [
-              'update "RelatedModel" set "test" = "test" - 10',
-              'where "RelatedModel"."rid" in',
-              '(select "JoinModel"."relatedId" from "JoinModel"',
-              'where "JoinModel"."ownerId" = 666)'
-            ].join(' ')
+            `update "RelatedModel" set "test" = "test" - 10 where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666))`
           );
         });
     });
@@ -1193,13 +1145,7 @@ describe('ManyToManyRelation', () => {
         .then(() => {
           expect(executedQueries).to.have.length(1);
           expect(executedQueries[0]).to.eql(
-            [
-              'update "RelatedModel" set "a" = \'str1\'',
-              'where "RelatedModel"."rid" in (select "JoinModel"."relatedId" from "JoinModel" where "JoinModel"."ownerId" = 666)',
-              'and "someColumn" = 100',
-              'and "gender" = \'male\'',
-              'and "thingy" is not null'
-            ].join(' ')
+            `update "RelatedModel" set "a" = 'str1' where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "someColumn" = 100 and "gender" = 'male' and "thingy" is not null)`
           );
         });
     });
@@ -1225,12 +1171,7 @@ describe('ManyToManyRelation', () => {
         expect(executedQueries[0]).to.equal(builder.toString());
         expect(executedQueries[0]).to.equal(builder.toSql());
         expect(executedQueries[0]).to.eql(
-          [
-            'delete from "RelatedModel"',
-            'where "RelatedModel"."rid" in (select "JoinModel"."relatedId" from "JoinModel" where "JoinModel"."ownerId" = 666)',
-            'and "gender" = \'male\'',
-            'and "thingy" is not null'
-          ].join(' ')
+          `delete from "RelatedModel" where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "gender" = 'male' and "thingy" is not null)`
         );
       });
     });
@@ -1254,12 +1195,7 @@ describe('ManyToManyRelation', () => {
         expect(executedQueries[0]).to.equal(builder.toString());
         expect(executedQueries[0]).to.equal(builder.toSql());
         expect(executedQueries[0]).to.eql(
-          [
-            'delete from "RelatedModel"',
-            'where ("RelatedModel"."cid","RelatedModel"."did") in (select "JoinModel"."relatedCId", "JoinModel"."relatedDId" from "JoinModel" where "JoinModel"."ownerAId" = 11 and "JoinModel"."ownerBId" = 22)',
-            'and "gender" = \'male\'',
-            'and "thingy" is not null'
-          ].join(' ')
+          `delete from "RelatedModel" where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."cid" = "JoinModel"."relatedCId" and "RelatedModel"."did" = "JoinModel"."relatedDId" where ("JoinModel"."ownerAId", "JoinModel"."ownerBId") in ((11, 22)) and "gender" = 'male' and "thingy" is not null)`
         );
       });
     });
@@ -1280,13 +1216,7 @@ describe('ManyToManyRelation', () => {
           expect(executedQueries).to.have.length(1);
           expect(result).to.eql({});
           expect(executedQueries[0]).to.eql(
-            [
-              'delete from "RelatedModel"',
-              'where "RelatedModel"."rid" in (select "JoinModel"."relatedId" from "JoinModel" where "JoinModel"."ownerId" = 666)',
-              'and "someColumn" = 100',
-              'and "gender" = \'male\'',
-              'and "thingy" is not null'
-            ].join(' ')
+            `delete from "RelatedModel" where "RelatedModel"."id" in (select "RelatedModel"."id" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "someColumn" = 100 and "gender" = 'male' and "thingy" is not null)`
           );
         });
     });
@@ -1504,14 +1434,7 @@ describe('ManyToManyRelation', () => {
         expect(executedQueries[0]).to.equal(builder.toString());
         expect(executedQueries[0]).to.equal(builder.toSql());
         expect(executedQueries[0]).to.eql(
-          [
-            'delete from "JoinModel"',
-            'where "JoinModel"."ownerId" = 666',
-            'and "JoinModel"."relatedId" in',
-            '(select "RelatedModel"."rid" from "RelatedModel"',
-            'where "code" in (55, 66, 77)',
-            'and "someColumn" = 100)'
-          ].join(' ')
+          `delete from "JoinModel" where "JoinModel"."relatedId" in (select "RelatedModel"."rid" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."rid" = "JoinModel"."relatedId" where "JoinModel"."ownerId" in (666) and "someColumn" = 100 and "code" in (55, 66, 77)) and "JoinModel"."ownerId" = 666`
         );
       });
     });
@@ -1534,15 +1457,7 @@ describe('ManyToManyRelation', () => {
         expect(executedQueries[0]).to.equal(builder.toString());
         expect(executedQueries[0]).to.equal(builder.toSql());
         expect(executedQueries[0]).to.eql(
-          [
-            'delete from "JoinModel"',
-            'where "JoinModel"."ownerAId" = 11',
-            'and "JoinModel"."ownerBId" = 22',
-            'and ("JoinModel"."relatedCId","JoinModel"."relatedDId") in',
-            '(select "RelatedModel"."cid", "RelatedModel"."did" from "RelatedModel"',
-            'where "code" in (55, 66, 77)',
-            'and "someColumn" = 100)'
-          ].join(' ')
+          `delete from "JoinModel" where ("JoinModel"."relatedCId","JoinModel"."relatedDId") in (select "RelatedModel"."cid", "RelatedModel"."did" from "RelatedModel" inner join "JoinModel" on "RelatedModel"."cid" = "JoinModel"."relatedCId" and "RelatedModel"."did" = "JoinModel"."relatedDId" where ("JoinModel"."ownerAId", "JoinModel"."ownerBId") in ((11, 22)) and "code" in (55, 66, 77) and "someColumn" = 100) and "JoinModel"."ownerAId" = 11 and "JoinModel"."ownerBId" = 22`
         );
       });
     });
